Consolidated integration tests
