+++ test/MLOps.NET.IntegrationTests/MLOps.NET.IntegrationTests.csproj
@@ -15,6 +15,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\MLOps.NET\MLOps.NET.csproj" />
+    <ProjectReference Include="..\MLOps.NET.Tests.Common\MLOps.NET.Tests.Common.csproj" />
   </ItemGroup>
 
 </Project>
