+++ test/MLOps.NET.SQLite.IntegrationTests/MLOps.NET.SQLite.IntegrationTests.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -20,6 +20,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\MLOps.NET.SQLite\MLOps.NET.SQLite.csproj" />
+    <ProjectReference Include="..\MLOps.NET.IntegrationTests\MLOps.NET.IntegrationTests.csproj" />
     <ProjectReference Include="..\MLOps.NET.Tests.Common\MLOps.NET.Tests.Common.csproj" />
   </ItemGroup>
